+++ libsq/driver/driver_test.go
@@ -368,6 +368,7 @@ var coreDrivers = []drivertype.Type{
 	drivertype.MySQL,
 	drivertype.ClickHouse,
 	drivertype.Oracle,
+	drivertype.DuckDB,
 	drivertype.CSV,
 	drivertype.TSV,
 	drivertype.XLSX,
@@ -381,6 +382,7 @@ var sqlDrivers = []drivertype.Type{
 	drivertype.MySQL,
 	drivertype.ClickHouse,
 	drivertype.Oracle,
+	drivertype.DuckDB,
 }
 
 // docDrivers is a slice of the doc driver types.